+++ common/src/main/java/dev/xylonity/companions/common/tesla/behaviour/dinamo/DinamoPulseBehaviour.java
@@ -7,6 +7,7 @@ import dev.xylonity.companions.common.tesla.ConnectionTarget;
 import dev.xylonity.companions.common.util.Util;
 import dev.xylonity.companions.common.util.interfaces.ITeslaGeneratorBehaviour;
 import dev.xylonity.companions.common.util.interfaces.ITeslaUtil;
+import dev.xylonity.companions.config.CompanionsConfig;
 import dev.xylonity.companions.registry.CompanionsEffects;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -90,7 +91,7 @@ public class DinamoPulseBehaviour implements ITeslaGeneratorBehaviour {
 
                     final Entity owner = sv.getEntity(dinamo.getOwnerUUID());
                     if (!Util.areEntitiesLinked(owner, victim) && !Util.areTeammates(owner, victim)) {
-                        victim.hurt(victim.level().damageSources().lightningBolt(), 7f);
+                        victim.hurt(victim.level().damageSources().lightningBolt(), (float) CompanionsConfig.ELECTRICITY_DAMAGE);
                         victim.addEffect(new MobEffectInstance(CompanionsEffects.ELECTROSHOCK.get(), 50, 0, false, true, true));
                     }
                 }